+++ AMQPSender/src/main/java/org/reliable/infrastructure/AMQPSender/Notifications.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 @JsonDeserialize(using = NotificationsDeserializer.class)
 public class Notifications {
-	private Integer idalert;
+	private Long idalert;
 	private String notificationType;
 	private String message;
 	private String priority;
@@ -17,7 +17,7 @@ public class Notifications {
 	public Notifications() {}
 	
 
-	public Notifications(Integer idalert, String notificationType, String message, String priority, List<String> tokens,
+	public Notifications(Long idalert, String notificationType, String message, String priority, List<String> tokens,
 																	List<String> phoneNumbers) {
 		this.idalert = idalert;
 		this.notificationType = notificationType;
@@ -29,11 +29,11 @@ public class Notifications {
 
 
 
-	public Integer getIdalert() {
+	public Long getIdalert() {
 		return idalert;
 	}
 
-	public void setIdalert(Integer idalert) {
+	public void setIdalert(Long idalert) {
 		this.idalert = idalert;
 	}
